Starting to test our RESTtful API itself
Some tests to POST (create), GET (list) and PUT (update) are already
written, but we have currenly only 49% of the `api` module tested. Let's
improve it in the next commits.

Just to remember that the api.py module (as in model.py) was extracted
from the bills software mentioned in the very first commit of this repo.