use python-dateutil 1.5 or 2.1
I've been running flask-restless with python-dateutil 2.1 in my project since I started, and I even have a test showing that timezone conversion is working correctly.

The only use of python-dateutil is the `dateutil.parse()` call and that doesn't appear to have changed between 1.5 and 2.1. It seems the major difference between 1.5 and 2.1 is that 2.1 is compatible with Python 3.