Add hybrid_properties to list of columns from helper
When performing a post to create a new instance we were not allowing the user to
set values for a hybrid_property since it was not in the set of defined columns.
However, a patch/put request worked just fine to update the attribute. This
change fixes it so you can set the hybrid_property attribute during a post
request.