Fixed `__repr__` functions to work with Python 2.6 and above
In a Python 2.6 repl an exception is thrown when viewing instances of
the Filter, SearchParameters, and OrderyBy classes. This is because
the call to `str.format` in the `__repr__` method of each omits the
indexes for the arguments, which is a feature added in Python 2.7.

From the Python docs [7.1.3. Format String Syntax][1], "Changed in
version 2.7: The positional argument specifiers can be omitted, so '{}
{}' is equivalent to '{0} {1}'."

[1]: http://docs.python.org/2/library/string.html#format-string-syntax