Makes exception available at top-level package
This makes `IllegalArgumentError` available to be imported directly from
`flask.ext.restless`. Before, in order to catch the exception, the user
would need to import it from `flask.ext.restless.manager`. This caused
an issue with Flask's import redirect for extensions as described at
mitsuhiko/flask#1347 in which the exception would not be catchable.