Allows Boolean expressions in search queries
Before, filters in search queries were joined using only conjunction or
only disjunction, as controlled by the `disjunction` parameter. Now we
allow arbitrary Boolean expressions of filter objects. By default, if no
ANDs or ORs are specified, the filters are assumed to be a conjunction.

This supercedes pull request #198 and fixes it in a different way.