Makes API compliant with JSON API specification.
Previously, the behavior of Flask-Restless was a bit arbitrary. Now we
force it to comply with a concrete (though still changing)
specification, which can be found at http://jsonapi.org/.

This is a (severely) backwards-incompatible change, as it changes which
API endpoints are exposed and the format of requests and responses.

This change also moves JSON API compliance tests to a convenient
distinct test module, `tests.test_jsonapi.py`, so that compliance with
the specification can be easily verified. These tests correspond to
version 1.0rc3 of the JSON API specification, which can be found in
commit json-api/json-api@a75f88d00634ff6dbb205f4e3625599a2c27a99c.

This change fixes (or at least makes it much easier to fix or much
easier to mark as "won't fix") quite a few issues.