Only remove the events if they have actually been listened.
These lines were put in place to fix a bug with Flask-SQLAlchemy adding listeners in a way which causes problems for a pure SQLAlchemy session. It seems that this is fixed in version 2.0 of Flask-SQLAlchemy, and the fix is now a bug. Adding the condition with "contains" means that this should now work on both older and newer version of Flask-SA (although it is untested on older versions).