Corrects Accept and Content-Type header handling.
The Accept headers now are not required, but if they are there, there
must be at least one instance with the JSON API media type,
`application/vnd.api+json`, without media type parameters.

The Content-Type headers are now only checked for `POST` and `PATCH`
requests, the only requests that actually have a body that can be
assigned a type. These headers are required for those requests.