Changes to class-based full document serialization
Serialization and deserialization is now done using serializer and
deserializer classes. The former has a `serialize()` and
`serialize_many()`, each of which produces a full JSON API document as a
Python dictionary. The latter class has a `deserialize()` method, which
receives a full JSON API document as input. Before, the serialization
functions sent and received only the primary data.

This commit also
- organizes the (de-)serialization code in a new `serialization`
  subpackage,
- moves various serialization code out of the `views` package and into
  the `serialization` package,
- updates the marshmallow example to the new JSON API format,
- allows the user to better generate serialization exceptions in custom
  serializers.