Removes dependencies on nose.
Well, not entirely. The `setup.py` module still specifies that the test
suite is the nose test collector and that the tests require
nose. However, this commit should allow the tests to run both with and
without nose (for example, with just `python setup.py test`). This is
intended to future-proof the code, since nose is no longer actively
maintained.