Skips PostgreSQL tests if creating tables fails
This adds a test skip if we are unable to create tables in a PostgreSQL
database. This change should have been incorporated in
a7b0814f1f03d13d1785471425f82bcbfb6b4136, which was accidentally merged
without a commit message. That commit should have had the following
commit message:

This commit removes the PostgreSQL testing from the other tests so that
it is easier to skip if desired. This also removes the test dependency
on `testing.postgresql` in favor of just creating the database in the
Travis configuration directly (using the technique recommended by Travis
itself).